Update README comprehensively. Also preprocessing will now recognise SVs that have already been processed and will skip them accordingly. Fixed bug where soft clip argument got mixed up with insert len. Cleaned up use of soft-threshold "tr" param.

+++ phylo_sv/preprocess/cmd.py
@@ -33,7 +33,7 @@ bam     = args.bam
 out     = args.out
 hd_cfg  = args.header
 mean_dp = float(args.mean_dp)
-sc_len  = float(args.sc_len)
+sc_len  = int(args.sc_len)
 max_cn  = int(args.max_cn)
 
 if __name__ == '__main__':
